requirements.txt
imutils==0.5.4
joblib==1.1.0
matplotlib==3.5.2
numpy==1.22.2
opencv_python==4.5.5.64
pandas==1.5.2
Pillow==9.3.0
scikit_image==0.19.3
scikit_learn==1.1.3
scipy==1.8.0
skimage==0.0
streamlit==1.15.1
tensorflow==2.8.0
